Fix MUI bug
UI renders properly, but Emscripten doesn't yet load properly
