+++ Lux/ControlPanel.js
@@ -14,7 +14,7 @@ import RadioGroup from '@mui/material/RadioGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
-import Box from '@mui/material/Box';
+import {Box} from '@mui/material';
 
 import { useTheme } from '@mui/material/styles';
 import { styled } from '@mui/system';
