+++ Lux/lux_react/src/components/ImagePortCanvas.jsx
@@ -3,7 +3,7 @@ import React, { useState, useRef, useEffect, useCallback } from 'react';
 function ImagePortCanvas( { width, height } ) {
   const canvasRef = useRef(null);
   const [isModuleReady, setModuleReady] = useState( false );
-  const [ frameTime, setFrameTime ] = useState( 0 );
+ // const [ frameTime, setFrameTime ] = useState( 0 );
 
   const handleMouseDown = useCallback(() => {
     if (window.Module) {
@@ -34,7 +34,6 @@ function ImagePortCanvas( { width, height } ) {
       const rect = canvasRef.current.getBoundingClientRect();
       const x = event.clientX - rect.left;
       const y = event.clientY - rect.top;
-      //console.log( "handleMouseMove x=" + x + " y=" + y + " width=" + width + " height=" + height );
       window.Module.mouse_move(x, y, width, height );
       window.Module.mouse_over( true );
     }
@@ -56,26 +55,26 @@ function ImagePortCanvas( { width, height } ) {
     const bufWidth  = window.Module.get_buf_width();
     const bufHeight = window.Module.get_buf_height();
     
-    let t1, t2, t3, t4;
-    t1 = performance.now();
+    //let t1, t2, t3, t4;
+    //t1 = performance.now();
     //console.log( "updateCanvas get_img_data time = " + (t1 - t0) );
     const imageData = new ImageData(
       new Uint8ClampedArray(imageDataArray.buffer, imageDataArray.byteOffset, imageDataArray.byteLength),
       bufWidth,
       bufHeight
     );
-    t2 = performance.now();
+    //t2 = performance.now();
     //console.log( "updateCanvas new ImageData time = " + (t2 - t1) );
     const imageBitmap = await createImageBitmap(imageData);
-    t3 = performance.now();
+    //t3 = performance.now();
     //console.log( "updateCanvas createImageBitmap time = " + (t3 - t2) );
     //console.log( "width=" + width + " height=" + height );
     ctx.drawImage(imageBitmap, 0, 0, width, height);
     //ctx.putImageData(imageData, 0, 0);
-    t4 = performance.now();
+    //t4 = performance.now();
     //console.log( "updateCanvas drawImage time = " + (t4 - t3) );
     //console.log( "frame time = " + (t4 - frameTime) );
-    setFrameTime( t4 );
+    //setFrameTime( t4 );
 
   }, [width, height]);
