+++ Lux/lux_react/src/components/InterfaceContainer.jsx
@@ -1,4 +1,4 @@
-import Box from "@mui/material/Box";
+import {Box} from "@mui/material";
 import React, { useEffect, useState } from "react";
 import ControlPanel from "./ControlPanel";
 import ImagePort from "./ImagePort";
