+++ Lux/lux_react/src/components/JenMultiDirection8.jsx
