+++ Lux/lux_react/src/components/JenSlider.jsx
