+++ Lux/lux_react/src/components/MediaController.jsx
@@ -1,15 +1,13 @@
 import React, { useState } from 'react';
-//import useEmscripten from './useEmscripten';
 import { ButtonGroup, Button, Tooltip } from '@mui/material';
-import RestartIcon from '@mui/icons-material/Replay'; // This is just an example icon for "restart"
-import FrameIcon from '@mui/icons-material/SkipNext'; // This is an example icon for "advance one frame"
+import RestartIcon from '@mui/icons-material/Replay';      // This is just an example icon for "restart"
+import FrameIcon from '@mui/icons-material/SkipNext';      // This is an example icon for "advance one frame"
 import PlayPauseIcon from '@mui/icons-material/PlayArrow'; // This is an example icon for "play"
-import PauseIcon from '@mui/icons-material/Pause'; // This is an example icon for "pause"
+import PauseIcon from '@mui/icons-material/Pause';         // This is an example icon for "pause"
 import WidgetContainer from './WidgetContainer';
 
 function MediaController( { panelSize } ) { 
   const [isRunning, setIsRunning] = useState(true);
-//  module = useEmscripten();
 
   const handleRestart = () => {
     if ( window.Module ) {
@@ -66,5 +64,3 @@ function MediaController( { panelSize } ) {
 };
 
 export default MediaController;
-
-
