+++ Lux/lux_react/src/components/WidgetContainer.jsx
