+++ Lux/lux_react/src/index.jsx
@@ -1,15 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import App from './App.jsx';
 
-
+/*
 var factory = require('./lux.js');
 
 factory().then((instance) => {
   window.Module = instance; // Assign to global variable
   window.Module.canvas = document.getElementById('imagePortCanvas');
+});
+*/
+
+import('./lux.js').then((instance) => {
+  window.module = instance; // Assign to global variable
+//  window.module.canvas = document.getElementById('imagePortCanvas');
 });
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -19,7 +24,4 @@ root.render(
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+
