+++ Lux/lux_react/vite.config.js
@@ -1,0 +1,18 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+    base: '',
+    plugins: [react()],
+    server: {    
+        // this ensures that the browser opens upon server start
+        open: true,
+        // this sets a default port to 3000  
+        port: 3000, 
+    },
+    resolve: {
+        alias: {
+          '@': '/src',
+        },
+      },
+});
