+++ Lux/web/components/SceneView.js
@@ -5,7 +5,7 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import Typography from '@mui/material/Typography';
-import Box from '@mui/material/Box';
+import {Box} from '@mui/material';
 
 function SceneView() {
   // This array contains our top-level labels for the accordion
